Move date picker logic into MedicalConsultation component
